feat: new DOM feature HandleLinkPressFeature
BREAKING CHANGE: handleLinkPressFeature has been replaced with
HandleLinkPressFeature, its instanciable counterpart. Also, the shape of
the sent message has changed, and is now an object with a variety of
useful information. Finally, this implementation uses the element
instance `href` field instead of the `href` attribute, which could contain
relative paths. It is thus more accurate.